requirements.txt
streamlit>=1.35.0
numpy==2.3.5
pandas==2.3.3
plotly>=6.0.0
websocket-client>=1.8.0
statsmodels>=0.14.0
scipy
scikit-learn
python-dateutil
pytz
tzdata
